Corrected a lot of bugs in the CUDA kernel related to dimensions: dim_x and dim_y, x and y. (These are tricky).
Added images of the observer map in png format.

Added print decorators and information.

Added toolbox in tools.py.

Added progress bar.

+++ Device_src/cuda_tools.cc
@@ -1,7 +1,9 @@
 #include <stdio.h>
 #include <math.h>
+#include <cuda.h>
 
-__device__ double get_data_index(int x, int y, int dim_x){
+
+__device__ int get_data_index(int x, int y, int dim_x){
 
  return x + y * dim_x ;
